indexer serialize FileID
